python/2019-10-17_Music_Bot/requirements.txt
discord.py[voice]==1.0
ffmpeg
youtube_dl
pafy